Move unit tests into the unit tests folder
